requirements.txt
flask
flask_cors
numpy
opencv-python
googletrans==3.1.0a0
requests
azure-mgmt-voiceservices
azure-cognitiveservices-speech
obs-websocket-py
pygame
soundfile
mutagen
keyboard
rich
tiktoken
mss
pillow
RealtimeTTS